+++ packages/terra-table/src/SelectableTableRows.jsx
@@ -22,6 +22,10 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
+   * A callback function to execute when the selected row loses focus. The first parameter is the event. The second parameter is the index of the row losing focus.
+   */
+  onBlur: PropTypes.func,
+  /**
    * The list of selected row indexes.
    */
   // eslint-disable-next-line react/forbid-prop-types
@@ -31,24 +35,77 @@ const propTypes = {
 const defaultProps = {
   disableUnselectedRows: false,
   onChange: undefined,
+  onBlur: undefined,
   selectedIndexes: [],
 };
+
+const rowsSelector = 'row-group';
 
 class SelectableTableRows extends React.Component {
   constructor(props) {
     super(props);
     this.handleOnChange = this.handleOnChange.bind(this);
+    this.handleOnBlur = this.handleOnBlur.bind(this);
+  }
+
+
+  componentDidMount() {
+    // Adds ARIA-Label for Pre-selected rows
+    const rowGroup = document.querySelector(`#${rowsSelector}`);
+    for (let rowIndex = 0; rowIndex < rowGroup.children.length; rowIndex += 1) {
+      const preselectedRow = rowGroup.children[rowIndex];
+      let ariaLabel = 'pre-selected row';
+      if ((preselectedRow.hasAttribute('aria-selected')) && (preselectedRow.getAttributeNode('aria-selected').value === 'true')) {
+        for (let columnIndex = 0; columnIndex < preselectedRow.children.length; columnIndex += 1) {
+          ariaLabel += ` ${preselectedRow.children[columnIndex].innerText}`;
+        }
+        rowGroup.children[rowIndex].setAttribute('aria-label', ariaLabel);
+      }
+    }
   }
 
   handleOnChange(event, index) {
     if (this.props.onChange) {
       this.props.onChange(event, index);
     }
+
+    if (!document.activeElement.hasAttribute('aria-label')) {
+      // Aria-Label is being populated with the cell contents for screen reader to read it as whole instead of cell value.
+      const ariaLabel = 'row selected';
+      document.activeElement.setAttribute('aria-label', ariaLabel);
+    }
+  }
+
+  handleOnBlur(event, index) {
+    if (this.props.onBlur) {
+      this.props.onBlur(event, index);
+    }
+  }
+
+  wrappedOnBlur(row, index) {
+    const initialOnBlur = row.props.onBlur;
+    // removes aria-label from unselected row
+    return (event) => {
+      if (row.props.isSelectable !== false) {
+        this.handleOnBlur(event, index);
+      }
+
+      const rowGroup = document.querySelector(`#${rowsSelector}`);
+      const previousRow = rowGroup.children[index];
+      if (previousRow.getAttributeNode('aria-selected').value === 'false') {
+        if (previousRow.hasAttribute('aria-label')) {
+          previousRow.removeAttribute('aria-label');
+        }
+      }
+
+      if (initialOnBlur) {
+        initialOnBlur(event);
+      }
+    };
   }
 
   wrappedOnClickForRow(row, index) {
     const initialOnClick = row.props.onClick;
-
     return (event) => {
       // The default isSelectable attribute is either undefined or true, unless the consumer specifies the row's isSelectable attribute as false.
       if (row.props.isSelectable !== false) {
@@ -78,7 +135,7 @@ class SelectableTableRows extends React.Component {
     };
   }
 
-  newPropsForRow(row, index, onClick, onKeyDown) {
+  newPropsForRow(row, index, onClick, onKeyDown, onBlur) {
     const isSelected = this.props.selectedIndexes.indexOf(index) >= 0;
     const newProps = { };
 
@@ -98,11 +155,12 @@ class SelectableTableRows extends React.Component {
     }
 
     // If selectable, add tabIndex on rows to navigate through keyboard tab key for selectable row and add
-    // onClick and onKeyDown functions.
+    // onClick, onKeyDown and onBlur functions.
     if (newProps.isSelectable) {
       newProps.tabIndex = '0';
       newProps.onClick = onClick;
       newProps.onKeyDown = onKeyDown;
+      newProps.onBlur = onBlur;
     }
 
     return newProps;
@@ -113,7 +171,8 @@ class SelectableTableRows extends React.Component {
       if (row.type !== TableHeader && row.type !== TableSubheader) {
         const wrappedOnClick = this.wrappedOnClickForRow(row, index);
         const wrappedOnKeyDown = this.wrappedOnKeyDownForRow(row, index);
-        const newProps = this.newPropsForRow(row, index, wrappedOnClick, wrappedOnKeyDown);
+        const wrappedOnBlur = this.wrappedOnBlur(row, index);
+        const newProps = this.newPropsForRow(row, index, wrappedOnClick, wrappedOnKeyDown, wrappedOnBlur);
         return React.cloneElement(row, newProps);
       }
       return row;
@@ -122,12 +181,12 @@ class SelectableTableRows extends React.Component {
 
   render() {
     const {
-      children, disableUnselectedRows, onChange, selectedIndexes, ...customProps
+      children, disableUnselectedRows, onChange, onBlur, selectedIndexes, ...customProps
     } = this.props;
     const clonedChildItems = this.clonedChildItems(children);
 
     return (
-      <TableRows {...customProps}>
+      <TableRows id={rowsSelector} {...customProps}>
         {clonedChildItems}
       </TableRows>
     );